requirements.txt
coverage==3.7.1
Django==1.8.3
django-envie==0.0.9
django-nose==1.4.2
django-rest-swagger==0.3.4
django-widget-tweaks==1.4.1
djangorestframework==3.3.1
djangorestframework-jwt==1.7.2
funcsigs==0.4
gunicorn==19.3.0
Markdown==2.6.4
mock==1.3.0
nose==1.3.7
pbr==1.8.1
psycopg2==2.6.1
PyJWT==1.4.0
PyYAML==3.11
requests==2.7.0
selenium==2.47.1
six==1.10.0
wheel==0.24.0
whitenoise==2.0.3
